refactor: cqrs 분리 및 코드 컨벤션 적용
